Tutorial/UseCase_A/model_serving/app/requirements_app.txt
mlflow~=2.3.1
numpy~=1.23.5
prefect~=2.11.1
python-dotenv~=0.21.0
torchvision
torch==2.0.*
flask
opencv-python
boto3
